zdbd: move auth commands to own file
