+++ zdbd/commands_auth.h
@@ -1,0 +1,5 @@
+#ifndef ZDB_COMMANDS_AUTH_H
+    #define ZDB_COMMANDS_AUTH_H
+
+    int command_auth(redis_client_t *client);
+#endif
